Database connected and a wrapper built
The new database class in the db.hpp file contains a 'run' method. This method allows for creating lambda functions that run on each individual element in the response. This is extremely useful. An example function I created is the loadCartProducts method. This calls the run method, then passes an external variable into the lambda function, where the results are processed and the external variable is updated.
